+++ x509-cert/src/certificate.rs
@@ -9,7 +9,10 @@ use der::{Decode, Enumerated, Error, ErrorKind, Sequence, Tag, ValueOrd};
 use spki::{AlgorithmIdentifierOwned, SubjectPublicKeyInfoOwned};
 
 #[cfg(feature = "pem")]
-use der::pem::PemLabel;
+use der::{
+    pem::{self, PemLabel},
+    DecodePem,
+};
 
 /// [`Profile`] allows the consumer of this crate to customize the behavior when parsing
 /// certificates.
@@ -212,3 +215,55 @@ impl<P: Profile> PemLabel for CertificateInner<P> {
 ///
 /// [RFC 6066]: https://datatracker.ietf.org/doc/html/rfc6066#section-10.1
 pub type PkiPath = Vec<Certificate>;
+
+#[cfg(feature = "pem")]
+impl<P: Profile> CertificateInner<P> {
+    /// Parse a chain of pem-encoded certificates from a slice.
+    ///
+    /// Returns the list of certificates.
+    pub fn load_pem_chain(mut input: &[u8]) -> Result<Vec<Self>, Error> {
+        fn find_boundary<T>(haystack: &[T], needle: &[T]) -> Option<usize>
+        where
+            for<'a> &'a [T]: PartialEq,
+        {
+            haystack
+                .windows(needle.len())
+                .position(|window| window == needle)
+        }
+
+        let mut certs = Vec::new();
+        let mut position: usize = 0;
+
+        let end_boundary = &b"-----END CERTIFICATE-----"[..];
+
+        // Strip the trailing whitespaces
+        loop {
+            if input.is_empty() {
+                break;
+            }
+            let last_pos = input.len() - 1;
+
+            match input.get(last_pos) {
+                Some(b'\r') | Some(b'\n') => {
+                    input = &input[..last_pos];
+                }
+                _ => break,
+            }
+        }
+
+        while position < input.len() - 1 {
+            let rest = &input[position..];
+            let end_pos = find_boundary(rest, end_boundary)
+                .ok_or(pem::Error::PostEncapsulationBoundary)?
+                + end_boundary.len();
+
+            let cert_buf = &rest[..end_pos];
+            let cert = Self::from_pem(cert_buf)?;
+            certs.push(cert);
+
+            position += end_pos;
+        }
+
+        Ok(certs)
+    }
+}